+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Issue #590 - Delete unused variables with obsolete tag 
    - "Remove global variables" settings split into separate "Remove global variables" and "Remove protected global variables" settings
      - `removeProtectedGlobalVariables` property added to the `defaultRemoveUnusedVariablesSetting` setting
+ - Issue #591 - Limit Enum Object Names to 30 Chars in AL File Wizard
  - Issue #592 - Remove redundant ApplicationArea adds unnecessary quotes 
 
  `alOutline.defaultRemoveUnusedVariablesSettings`: default settings for the RemoveUnusedVariables command when run by the Code Cleanup, these properties can be set:
@@ -14,6 +15,7 @@
 Thank you
  - fvet and TKapitan for testing and adding comments to the issue #535
  - ThorstenEngelsGOB for reporting issue #590
+ - mjmatthiesen for reporting issue #591
  - dannoe for reporting issue #592
 
 ## 13.66.0